fix exception in cms due to missing THIRD_PARTY_AUTH_BACKENDS
