+++ tahoe_idp/tests/test_apps.py
@@ -10,23 +10,23 @@ def test_app_config():
         'settings_config': {
             'cms.djangoapp': {
                 'production': {  # Needs to be production settings, otherwise devstack will override it
-                    'relative_path': 'settings.production'
+                    'relative_path': 'settings.cms_production',  # Use cms-settings
                 },
             },
             'lms.djangoapp': {
                 'production': {  # Needs to be production settings, otherwise devstack will override it
-                    'relative_path': 'settings.production'
+                    'relative_path': 'settings.lms_production',  # Use lms-settings
                 },
-            }
+            },
         },
         'url_config': {
             'cms.djangoapp': {
                 'namespace': 'tahoe_idp',
-                'app_name': 'tahoe_idp'
+                'app_name': 'tahoe_idp',
             },
             'lms.djangoapp': {
                 'namespace': 'tahoe_idp',
-                'app_name': 'tahoe_idp'
-            }
-        }
+                'app_name': 'tahoe_idp',
+            },
+        },
     }, 'Should initiate the app as an Open edX plugin.'
